app/requirement.txt
fastapi
uvicorn
pydantic
sqlalchemy
python-jose
passlib[bcrypt]
pymysql
psycopg2-binary
python-dotenv
argon2-cffi
python-multipart
python-jose[cryptography]
python-multipart